+++ R/globals.R
@@ -1,1 +1,1 @@
-globalVariables(c('prov_code','1999_nickname','.'))
+globalVariables(c('prov_code','1999_nickname','.','dia_sub_group','freq', 'prov_scode'))