+++ tests/testthat/test-regioncode.R
@@ -5,7 +5,7 @@ test_that("multiplication works", {
                           year_to = 1989)), )
 })
 
-region_table[which(is.na(regioncode(data_input = corruption$prefecture_id,
+region_data[which(is.na(regioncode(data_input = corruption$prefecture_id,
                  convert_to = "name",
                  year_from = 2019,
                  year_to = 1989))),]