+++ NEWS.md
@@ -3,6 +3,7 @@
 - Bug fix
   - present Shaanxi, tibet, macao, hongkong, inner_mongolia correctly
   - full name of "Ning Xia"
+  - Correct name of Xinjiang
 + Add new function `rank` 
 
 # regioncode 0.1.1